refactor: 마이그레이션 단계별 모듈 구조 정리
