+++ app/src/step6/useclient-minimizer.cjs
